Break project into class files
